feat: extract navigation handlers
